requirements.txt
pycryptodome
gevent
msgpack
protobuf